# src/main/resources/api-spec/oas-wigo4it-koppelvak-lo-2026.030.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LO Behandeld API",
    "version": "v1"
  },
  "paths": {
    "/api/v{version}/LOBehandeld": {
      "post": {
        "tags": [
          "LOBehandeld"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ZaakLOBehandeld"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaakLOBehandeld"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ZaakLOBehandeld"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ZaakLOBehandeld"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLOBehandeld"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLOBehandeld"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseLOBehandeld"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Adres": {
        "required": [
          "soortAdres"
        ],
        "type": "object",
        "oneOf": [
          {
            "$ref": "#/components/schemas/NederlandsPostbusadres"
          },
          {
            "$ref": "#/components/schemas/NederlandsStraatadres"
          },
          {
            "$ref": "#/components/schemas/StraatadresBuitenland"
          }
        ],
        "properties": {
          "soortAdres": {
            "$ref": "#/components/schemas/SoortAdres"
          }
        },
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "soortAdres",
          "mapping": {
            "NederlandsPostbusadres": "#/components/schemas/NederlandsPostbusadres",
            "NederlandsStraatadres": "#/components/schemas/NederlandsStraatadres",
            "StraatadresBuitenland": "#/components/schemas/StraatadresBuitenland"
          }
        }
      },
      "Adreshouding": {
        "required": [
          "adres",
          "begindatum",
          "codeFunctieAdres"
        ],
        "type": "object",
        "properties": {
          "adres": {
            "$ref": "#/components/schemas/Adres"
          },
          "codeFunctieAdres": {
            "$ref": "#/components/schemas/AdreshoudingCodeFunctieAdres"
          },
          "begindatum": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "AdreshoudingCodeFunctieAdres": {
        "enum": [
          "A",
          "B",
          "C",
          "L",
          "V",
          "W"
        ],
        "type": "string",
        "description": "A = A, B = B, C = C, L = L, V = V, W = W"
      },
...
      "Besluit": {
        "required": [
          "dienst"
        ],
        "type": "object",
        "properties": {
          "dienst": {
            "$ref": "#/components/schemas/Dienst"
          }
        },
        "additionalProperties": false
      },
...
      "Betrokkene": {
        "required": [
          "naam"
        ],
        "type": "object",
        "properties": {
          "naam": {
            "type": "string",
            "description": "De feitelijke naam van de organisatorische eenheid.Bron: RGBZ2.0"
          },
          "omschrijving": {
            "type": "string",
            "description": "Een omschrijving van de organisatorische eenheid.Bron: RGBZ2.0",
            "nullable": true
          },
          "rol": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rol"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Bewindvoerder": {
        "required": [
          "rolOrganisatie",
          "vestiging"
        ],
        "type": "object",
        "properties": {
          "rolOrganisatie": {
            "$ref": "#/components/schemas/BewindvoerderRolOrganisatie"
          },
          "vestiging": {
            "$ref": "#/components/schemas/Vestiging"
          }
        },
        "additionalProperties": false
      },
      "BewindvoerderRolOrganisatie": {
        "enum": [
          "AR",
          "C",
          "B"
        ],
        "type": "string",
        "description": "AR = AR, C = C, B = B"
      },
...
      "Contactgegevens": {
        "type": "object",
        "properties": {
          "codeTypeTelefoonnummer": {
            "$ref": "#/components/schemas/ContactgegevensCodeTypeTelefoonnummer"
          },
          "eMailadres": {
            "maxLength": 70,
            "pattern": "^[A-Za-z0-9_-]+([.][A-Za-z0-9_-]+)*@[A-Za-z0-9_-]+([.][A-Za-z0-9_-]+)*[.][A-Za-z0-9]+$",
            "type": "string",
            "description": "Het adres waaronder een PERSOON per elektronische post bereikbaar is.",
            "nullable": true
          },
          "indicatieGeheimTelefoonnummer": {
            "$ref": "#/components/schemas/JaNee"
          },
          "indicatieVoorkeurTelefoonnummer": {
            "$ref": "#/components/schemas/JaNee"
          },
          "telefoonLandnummer": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "description": "Het telefoon landnummer waaronder de PERSOON bereikbaar is.Norminstantie: ITU",
            "format": "int32",
            "nullable": true
          },
          "telefoonnummer": {
            "pattern": "^0[1-9][0-9]{8}$",
            "type": "string",
            "description": "Het telefoonnummer waaronder een PERSOON bereikbaar is.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Dienst": {
        "required": [
          "begindatum",
          "einddatum"
        ],
        "type": "object",
        "properties": {
          "begindatum": {
            "type": "string",
            "format": "date"
          },
          "einddatum": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "Geheim": {
        "enum": [
          "0",
          "1",
          "2",
          "3",
          "4",
          "5",
          "6",
          "7"
        ],
        "type": "string",
        "description": "Code die aangeeft dat de persoonsgegevens, volgens Basisregistratie Personen, wel of niet aan een andere organisatie (derden, kerken) verstrekt mogen worden.\n\n 0 = GeenRestricties, 1 = NietDerdenVerbindendVoorschrift, 2 = NietKerken, 3 = NietVrijeDerden, 4 = NietDerdenVerbindendVoorschriftEnKerken, 5 = NietDerdenVerbindendVoorschriftEnDerden, 6 = NietKerkenEnDerden, 7 = NietDerdenVerbindendVoorschriftEnKerkenEnDerden"
      },
      "Geslachtsaanduiding": {
        "enum": [
          "0",
          "1",
          "2"
        ],
        "type": "string",
        "description": "0 = Onbekend, 1 = Man, 2 = Vrouw"
      },
      "Huishouding": {
        "required": [
          "aanvrager"
        ],
        "type": "object",
        "properties": {
          "aanvrager": {
            "$ref": "#/components/schemas/Persoon"
          },
          "inAanmerkingTeNemenVermogen": {
            "pattern": "^-?[0-9]+((;|,|\\\\.)[0-9]{2}|)$",
            "type": "string",
            "description": "Totaal van het in aanmerking te nemen vermogen van leden gezamenlijke huishouding plus de vermogens van de ten laste komende kinderen die op zelfde adres wonen.",
            "nullable": true
          },
          "leefsituatie": {
            "$ref": "#/components/schemas/Leefsituatie"
          },
          "medeaanvrager": {
            "$ref": "#/components/schemas/Persoon"
          },
          "uitbetalingUitkering": {
            "$ref": "#/components/schemas/UitbetalingUitkering"
          }
        },
        "additionalProperties": false
      },
      "Huisvesting": {
        "required": [
          "codeRelatieBewonerTotHuisvesting",
          "indicatieZelfstandigeWoonruimte"
        ],
        "type": "object",
        "properties": {
          "aantalDagenPerWeekInInstellingOfOpvang": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "description": "Het aantal dagen per week dat men in een instelling of opvang verblijft. Dit attribuut moet zijn ingevuld als de CodeRelatieBewonerTotHuisvesting aangeeft dat men in een instelling zit.",
            "format": "int32",
            "nullable": true
          },
          "codeRelatieBewonerTotHuisvesting": {
            "$ref": "#/components/schemas/HuisvestingCodeRelatieBewonerTotHuisvesting"
          },
          "indicatieHuur": {
            "$ref": "#/components/schemas/IndicatieHuur"
          },
          "indicatieInDetentie": {
            "$ref": "#/components/schemas/JaNee"
          },
          "indicatieInstellingVolledigVerzorgd": {
            "$ref": "#/components/schemas/JaNee"
          },
          "indicatieZelfstandigeWoonruimte": {
            "$ref": "#/components/schemas/JaNee"
          },
          "toelichtingHuisvesting": {
            "type": "string",
            "description": "Toelichting op de woonsituatie van de PERSOON betreffende de huisvesting.Bron: SGR 16.0",
            "nullable": true
          },
          "toelichtingRelatieBewonerTotHuisvesting": {
            "type": "string",
            "description": "Toelichting op de hoedanigheid waarin de PERSOON de woning bewoont.Norminstantie: SGR",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "IngangBijstandsuitkering": {
        "required": [
          "datumMeldingBijGemeente"
        ],
        "type": "object",
        "properties": {
          "afwijkendeIngangsdatum": {
            "$ref": "#/components/schemas/JaNee"
          },
          "datumMeldingBijGemeente": {
            "type": "string",
            "description": "Indien er direct een aanvraag is gedaan dan is de melddatum gelijk aan datum waarop aanvraag is ontvangen.",
            "format": "date"
          },
          "gewensteStartdatumUitkering": {
            "type": "string",
            "description": "Gewenste datum waarop uitkering ingaat",
            "format": "date",
            "nullable": true
          },
          "redenAfwijkendestartdatum": {
            "$ref": "#/components/schemas/RedenAfwijkendeStartdatum"
          }
        },
        "additionalProperties": false
      },
...
      "JaNee": {
        "enum": [
          "1",
          "2"
        ],
        "type": "string",
        "description": "1 = Ja, 2 = Nee"
      },
...
      "LOBehandeld": {
        "required": [
          "aanvraagdatum",
          "aanvraagid",
          "codeOntvangendeGemeente",
          "huishouding",
          "ingangBijstandsuitkering",
          "redenAanvraagLevensonderhoud"
        ],
        "type": "object",
        "properties": {
          "aanvraagdatum": {
            "type": "string",
            "description": "De datum waarop de aanvraag is ingediend.Doorgaans is dit de verzenddatum, echter, de aanvraag kan leiden tot een compleetheid- en juistheidtoets. Dit kan betekenen dat de aanvraag pas later wordt gecompleteerd.",
            "format": "date"
          },
          "aanvraagid": {
            "pattern": "^(\\{{0,1}([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12}\\}{0,1})$",
            "type": "string"
          },
          "besluit": {
            "$ref": "#/components/schemas/Besluit"
          },
          "codeOntvangendeGemeente": {
            "type": "string",
            "description": "De code van de gemeente, die de aanvraag ontvangt en verwerkt."
          },
          "huishouding": {
            "$ref": "#/components/schemas/Huishouding"
          },
          "ingangBijstandsuitkering": {
            "$ref": "#/components/schemas/IngangBijstandsuitkering"
          },
          "redenAanvraagLevensonderhoud": {
            "$ref": "#/components/schemas/RedenAanvraagLevensonderhoud"
          }
        },
        "additionalProperties": false
      },
      "Leefsituatie": {
        "enum": [
          "1",
          "2",
          "3",
          "4",
          "5",
          "6"
        ],
        "type": "string",
        "description": "1 = InstitutioneelHuishouden, 2 = Eenoudergezin, 3 = Alleenstaand, 4 = VastePartnersKinderen, 5 = VastePartnersGeenKinderen, 6 = OverigHuishouden"
      },
      "Medebewoner": {
        "required": [
          "bewoner",
          "indicatieKostendeler",
          "ingangsdatumRelatie",
          "relatietype"
        ],
        "type": "object",
        "properties": {
          "bewoner": {
            "$ref": "#/components/schemas/Bewoner"
          },
          "indicatieKostendeler": {
            "$ref": "#/components/schemas/JaNee"
          },
          "indicatieZorgbehoefte": {
            "$ref": "#/components/schemas/JaNee"
          },
          "relatietype": {
            "$ref": "#/components/schemas/MedebewonerRelatietype"
          },
          "ingangsdatumRelatie": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "Naamgebruik": {
        "enum": [
          "1",
          "2",
          "3",
          "4"
        ],
        "type": "string",
        "description": "Een aanduiding voor de wijze van aanschrijving van de NATUURLIJK PERSOON.\n\n 1 = EigenNaam, 2 = ExEchtgenoot, 3 = ExEchtgenootEigenNaam, 4 = EigenNaamExEchtgenoot"
      },
...
      "OuderKind": {
        "required": [
          "indicatieCoOuder",
          "indicatieKinderbijslag",
          "indicatieOnderhoudsplicht",
          "indicatieThuiswonend",
          "kind",
          "ouderKindRelatietype"
        ],
        "type": "object",
        "properties": {
          "indicatieCoOuder": {
            "$ref": "#/components/schemas/OuderKindIndicatieCoOuder"
          },
          "indicatieKinderbijslag": {
            "$ref": "#/components/schemas/JaNee"
          },
          "indicatieOnderhoudsplicht": {
            "$ref": "#/components/schemas/JaNee"
          },
          "indicatieThuiswonend": {
            "$ref": "#/components/schemas/JaNee"
          },
          "kind": {
            "$ref": "#/components/schemas/Kind"
          },
          "ouderKindRelatietype": {
            "$ref": "#/components/schemas/OuderKindRelatietype"
          }
        },
        "additionalProperties": false
      },
...
      "Partner": {
        "required": [
          "ingangsdatumRelatie",
          "partnerRelatietype"
        ],
        "type": "object",
        "properties": {
          "partnerRelatietype": {
            "$ref": "#/components/schemas/PartnerRelatietype"
          },
          "ingangsdatumRelatie": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "Persoon": {
        "required": [
          "burgerservicenummer",
          "codeBrpGegevensGeheim",
          "geboortedatum",
          "geslachtsaanduiding",
          "geslachtsnaamstam",
          "naamgebruik",
          "nationaliteit",
          "voorlettersAanschrijving"
        ],
        "type": "object",
        "properties": {
          "adreshouding": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Adreshouding"
            },
            "nullable": true
          },
          "aNummer": {
            "pattern": "^[1-9][0-9]{9}$",
            "type": "string",
            "description": "Het administratienummer van een PERSOON binnen de gemeentelijke basisadministratie (GBA), zoals bedoeld als in artikel 50 van de Wet GBA.Norminstantie: GBA",
            "nullable": true
          },
          "bewindvoerder": {
            "$ref": "#/components/schemas/Bewindvoerder"
          },
          "burgerservicenummer": {
            "pattern": "^[0-9]{9}$",
            "type": "string",
            "description": "Een uniek identificerend nummer van een PERSOON, die ingeschreven staat in de Gemeentelijke Basisadministratie Persoonsgegevens (GBA) of de Registratie Niet-ingezetenen (RNI).Bron: SGROpmerking:Elk nummer moet voldoen aan de bij de voorwaarden beschreven proef.Voorwaarden waarde BSN:  Indien het nummer per cijfer wordt aangeduid als (s0 s1 s2 s3 s4 s5 s6 s7 s8), dan is de volgende voorwaarde te controleren: (9 x s0) + (8 x s1) + (7 x s2) + (6 x s3) + (5 x s4) + (4 x s5) + (3 x s6) + (2 x s7) - (1 x s8) is deelbaar door 11Toelichting:  Het Burgerservicenummer onderscheidt zich van het Sofi-nummer voornamelijk voor wat betreft het bereik ervan en de wijze waarop wettelijk is vastgelegd wat met behulp van het Burgerservicenummer mag gebeuren.  De gebruikers van het Burgerservicenummer bestaan uit overheidsorganen alsmede niet-overheidsorganen. Omdat het niet duidelijk is wat de criteria zijn om niet-overheidsorganen wel of niet toe te laten tot het gebruik van het Burgerservicenummer is besloten om de gebruikersaspecten niet in de definitie op te nemen.De oorspronkelijke naamgeving en definitie van dit gegevenselement conform de (authentieke) bron:Elementnaam:    BurgerservicenummerDefinitie:    Elk ingeschreven persoon heeft een BSN, een nummer dat de persoon uniek identificeert in alle overheidsadministraties.Norminstantie:    GBA"
          },
          "contactgegevens": {
            "$ref": "#/components/schemas/Contactgegevens"
          },
          "geboortedatum": {
            "type": "string",
            "description": "De datum van geboorte van een PERSOON.Bron: NEN 1888:2002 - Geboortedatum (overgenomen zoals in SGR)Opmerking:Deze datum staat gewoonlijk vermeld op de geboorteakte dan wel een ander brondocument.Een GEBOORTEDATUM behoort te worden vastgesteld in verband met zijn rechtsgeldige betekenis, in het bijzonder voor de instellingen die minimaal een SZ-WET uitvoeren.",
            "format": "date"
          },
          "kinderen": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OuderKind"
            },
            "nullable": true
          },
          "medebewoners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Medebewoner"
            },
            "nullable": true
          },
          "partner": {
            "$ref": "#/components/schemas/Partner"
          },
          "profiel": {
            "$ref": "#/components/schemas/Profiel"
          },
          "verblijfstitel": {
            "$ref": "#/components/schemas/Verblijfstitel"
          },
          "huisvesting": {
            "$ref": "#/components/schemas/Huisvesting"
          },
          "geslachtsaanduiding": {
            "$ref": "#/components/schemas/Geslachtsaanduiding"
          },
          "geslachtsnaamstam": {
            "pattern": "^[\\p{L}\\p{M}\\-'\\.,\\/\\s]{1,40}$",
            "type": "string",
            "description": "De stam van de geslachtsnaam. Deze is ontdaan van voorvoegsels en titels."
          },
          "voorlettersAanschrijving": {
            "pattern": "^[\\p{L}. ]*$",
            "type": "string",
            "description": "De voorletters waarmee een persoon aangeschreven wil worden."
          },
          "voornamen": {
            "pattern": "^[\\p{L}\\p{M}\\-'\\.,\\s]{1,200}$",
            "type": "string",
            "description": "De samenvoeging van alle exemplaren van voornaam van een persoon.",
            "nullable": true
          },
          "voorvoegsel": {
            "pattern": "^[\\p{L}\\p{M}\\-'\\.,\\s]{1,10}$",
            "type": "string",
            "description": "Voorvoegsel behorende bij de stam van de geslachtsnaam.",
            "nullable": true
          },
          "naamgebruik": {
            "$ref": "#/components/schemas/Naamgebruik"
          },
          "geslachtsnaamPartner": {
            "pattern": "^[\\p{L}\\p{M}\\-'\\.,\\s]{1,51}$",
            "type": "string",
            "nullable": true
          },
          "codeBrpGegevensGeheim": {
            "$ref": "#/components/schemas/Geheim"
          },
          "nationaliteit": {
            "maxItems": 2,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CodeNationaliteit"
            },
            "description": "Codes die de Nationaliteiten van een PERSOON aangeeft."
          }
        },
        "additionalProperties": false
      },
...
      "Profiel": {
        "type": "object",
        "properties": {
          "inkomstenverhoudingen": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Inkomstenverhouding"
            },
            "nullable": true
          },
          "indNietRechthebbend": {
            "$ref": "#/components/schemas/JaNee"
          }
        },
        "additionalProperties": false
      },
      "RedenAanvraagLevensonderhoud": {
        "type": "object",
        "properties": {
          "onvoldoendeInkomen": {
            "$ref": "#/components/schemas/RedenAanvraagLevensonderhoudOnvoldoendeInkomen"
          },
          "teLaagInkomen": {
            "$ref": "#/components/schemas/RedenAanvraagLevensonderhoudTeLaagInkomen"
          },
          "uitkeringStopgezet": {
            "$ref": "#/components/schemas/RedenAanvraagLevensonderhoudUitkeringStopgezet"
          },
          "verblijfstatus": {
            "$ref": "#/components/schemas/RedenAanvraagLevensonderhoudVerblijfstatus"
          },
          "wijzigingGezin": {
            "$ref": "#/components/schemas/RedenAanvraagLevensonderhoudWijzigingGezin"
          },
          "zelfstandige": {
            "$ref": "#/components/schemas/RedenAanvraagLevensonderhoudZelfstandige"
          }
        },
        "additionalProperties": false
      },
      "RedenAanvraagLevensonderhoudOnvoldoendeInkomen": {
        "enum": [
          "1",
          "2",
          "3",
          "4",
          "5",
          "6",
          "7",
          "8",
          "9",
          "10",
          "11"
        ],
        "type": "string",
        "description": "1 = IkHebGeenWerkMeer, 2 = IkHebWelInkomstenMaarTeLaag, 3 = UitkeringGeenBijstandStopgezet, 4 = BijstandsuitkeringStopgezet, 5 = BijstandsaanvraagAfgewezenOfNietInBehandeling, 6 = AlimentatieGestoptOfVerlaagd, 7 = StudiefinancieringGestopt, 8 = InStakingOfUitgeslotenVoorWerk, 9 = VanSpaargeldGeleefd, 10 = VrijNaGevangenisstraf, 11 = AndereReden"
      },
      "RedenAanvraagLevensonderhoudTeLaagInkomen": {
        "enum": [
          "1",
          "2",
          "3",
          "4"
        ],
        "type": "string",
        "description": "1 = WWUitkeringTeLaag, 2 = ZWUitkeringTeLaag, 3 = TeLaagInkomenUitParttimeWerk, 4 = AanvullingOpAndereInkomsten"
      },
      "RedenAanvraagLevensonderhoudUitkeringStopgezet": {
        "enum": [
          "1",
          "2",
          "3",
          "4",
          "5"
        ],
        "type": "string",
        "description": "1 = EindeWWUitkering, 2 = EindeZWUitkering, 3 = EindeWGAUitkering, 4 = EindeWAOIVAUitkering, 5 = EindeOverigeUitkering"
      },
      "RedenAanvraagLevensonderhoudVerblijfstatus": {
        "enum": [
          "1",
          "2"
        ],
        "type": "string",
        "description": "1 = VerblijfsvergunningGekregen, 2 = VerblijfsvergunningGekregenEnVertrekUitAsielzoekerscentrum"
      },
      "RedenAanvraagLevensonderhoudWijzigingGezin": {
        "enum": [
          "1",
          "2",
          "3"
        ],
        "type": "string",
        "description": "1 = EigenUitkeringMaarNuSamenAanvraagMetPartner, 2 = RelatieVerbroken, 3 = PartnerOverleden"
      },
      "RedenAanvraagLevensonderhoudZelfstandige": {
        "enum": [
          "1",
          "2",
          "3"
        ],
        "type": "string",
        "description": "1 = GestoptMetEigenBedrijf, 2 = EigenBedrijfOnvoldoendeInkomen, 3 = EigenBedrijfTijdelijkGeslotenDoorGemeente"
      },
      "RedenAfwijkendeStartdatum": {
        "required": [
          "redenAfwijkingAanwezig"
        ],
        "type": "object",
        "properties": {
          "aanvraagdatumDigid": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "ontvangstdatumBeslissingInstantie": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "redenafwijkendestartdatumtype": {
            "$ref": "#/components/schemas/Redenafwijkendestartdatumtype"
          },
          "redenAfwijkingAanwezig": {
            "$ref": "#/components/schemas/JaNee"
          },
          "specificatieBijzondereRedenLaterIngaanStartdatumAanvraag": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Redenafwijkendestartdatumtype": {
        "enum": [
          "1",
          "2",
          "3",
          "4"
        ],
        "type": "string",
        "description": "1 = WachtenBeslissingInstantie, 2 = WachtenDigid, 3 = OpnameInstelling, 4 = AndereRedenAfwijkendeStartdatum"
      },
      "ResponseLOBehandeld": {
        "required": [
          "berichtId",
          "responseId"
        ],
        "type": "object",
        "properties": {
          "berichtId": {
            "type": "string",
            "description": "maakt een response aan met de minimale gegevens",
            "nullable": true
          },
          "responseId": {
            "type": "string",
            "description": "identificerend gegeven van verwerking door Wigo4it",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "datum tijd waarop verweking is afgerond",
            "nullable": true
          },
          "foutcode": {
            "type": "string",
            "description": "foutcode, indien er een fout is opgetreden",
            "nullable": true
          },
          "foutomschrijving": {
            "type": "string",
            "description": "korte functionele omschrijving van fout",
            "nullable": true
          },
          "details": {
            "type": "string",
            "description": "details van fout, indien aanwezig",
            "nullable": true
          },
          "fouten": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "lijst van individuele validatiefouten, indien aanwezig",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "response die altijd gegeven wordt"
      },
      "Rol": {
        "required": [
          "rolomschrijving"
        ],
        "type": "object",
        "properties": {
          "rolomschrijving": {
            "type": "string",
            "description": "Algemeen gehanteerde benaming van de aard van de ROL"
          }
        },
        "additionalProperties": false
      },
...
      "UitbetalingUitkering": {
        "required": [
          "soortUitbetaling"
        ],
        "type": "object",
        "properties": {
          "aanDerdeReden": {
            "type": "string",
            "description": "De reden waarom er aan een derde uitbetaald moet worden. Een derde is een bewindvoerder of een gemachtigde zoals een curator, een organisatie voor matschappelijke opvang of een famililid.",
            "nullable": true
          },
          "bankrekening": {
            "pattern": "^[A-Z]{2}[0-9]{2}[A-Z0-9]{11,30}$",
            "type": "string",
            "nullable": true
          },
          "bankrekeningDerde": {
            "pattern": "^[A-Z]{2}[0-9]{2}[A-Z0-9]{11,30}$",
            "type": "string",
            "description": "IBAN nummer van derde. Een derde is een bewindvoerder of een gemachtigde zoals een curator, een organisatie voor maatschappelijke opvang of een famililid.",
            "nullable": true
          },
          "bankrekeningPartner": {
            "pattern": "^[A-Z]{2}[0-9]{2}[A-Z0-9]{11,30}$",
            "type": "string",
            "description": "IBAN nummer van partner aanvrager",
            "nullable": true
          },
          "rekeninghouder": {
            "type": "string",
            "description": "Naam rekeninghouder.",
            "nullable": true
          },
          "rekeninghouderDerde": {
            "type": "string",
            "description": "Naam rekeninghouder derde. Een derde is een bewindvoerder of een gemachtigde zoals een curator, een organisatie voor maatschappelijke opvang of een famililid.",
            "nullable": true
          },
          "rekeninghouderPartner": {
            "type": "string",
            "description": "Naam rekeninghouder van partner aanvrager",
            "nullable": true
          },
          "soortUitbetaling": {
            "$ref": "#/components/schemas/UitbetalingUitkeringSoortUitbetaling"
          }
        },
        "additionalProperties": false
      },
      "UitbetalingUitkeringSoortUitbetaling": {
        "enum": [
          "01",
          "02",
          "03",
          "04",
          "05"
        ],
        "type": "string",
        "description": "01 = Klant, 02 = Partner, 03 = Gesplitst, 04 = Gemachtigde, 05 = Bewindvoerder"
      },
...
      "Verblijfstitel": {
        "type": "object",
        "properties": {
          "codeVerblijfstitel": {
            "$ref": "#/components/schemas/VerblijfstitelCodeVerblijfstitel"
          },
          "datumAanvangVerblijfstitel": {
            "type": "string",
            "description": "De datum van de eerste dag waarop de VERBLIJFSTITEL van toepassing is.Norminstantie: UGR",
            "format": "date",
            "nullable": true
          },
          "datumEindeVerblijfstitel": {
            "type": "string",
            "description": "De datum van de laatste dag waarop de VERBLIJFSTITEL van toepassing is.Norminstantie: UGR",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ZaakLOBehandeld": {
        "required": [
          "identificatie",
          "loBehandeld"
        ],
        "type": "object",
        "properties": {
          "betrokkenen": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Betrokkene"
            },
            "nullable": true
          },
          "identificatie": {
            "type": "string",
            "description": "De unieke identificatie van de ZAAK binnen de organisatie die verantwoordelijk is voor de behandeling van de ZAAK."
          },
          "loBehandeld": {
            "$ref": "#/components/schemas/LOBehandeld"
          },
          "url": {
            "type": "string",
            "description": "URL-referentie naar dit object. Dit is de unieke identificatie en locatie van dit object.",
            "nullable": true
          },
          "uuid": {
            "pattern": "^(\\\\{{0,1}([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12}\\\\}{0,1})$",
            "type": "string",
            "description": "De unieke identificatie van de ZAAK binnen de organisatie die verantwoordelijk is voor de behandeling van de ZAAK.",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}